moved xpls/version to internal/version moved folded local/remote version comments into errors returned to client
